livekit-testing/requirements.txt
livekit-agents>=1.3.0
livekit-plugins-silero>=0.7.0
python-dotenv>=1.0.0
